+++ packages/orq-rc/RELEASES.md
@@ -10989,3 +10989,13 @@ Based on:
 - [typescript v4.12.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.45 - packages/orq-rc
+
+## 2026-07-21 08:36:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.46 - packages/orq-rc